Refactor profiler into separate module
