+++ samples/apps/tinyRA/tinyra/tui.py
@@ -47,7 +47,7 @@ from autogen.coding.func_with_reqs import FunctionWithRequirements
 from .tools import Tool, InvalidToolError
 from .exceptions import ChatMessageError, ToolUpdateError, SubprocessError
 from .app_config import AppConfiguration
-
+from .profiler import Profiler, State, StateSpace
 
 APP_CONFIG = AppConfiguration()
 APP_CONFIG.initialize()
@@ -702,431 +702,12 @@ class AgentMessage:
         return f"{self.role}:\n{self.content}"
 
 
-@dataclass
-class State:
-
-    name: str
-    description: str
-    tags: List[str]
-
-    def __str__(self):
-        return f"{self.name}"
-
-    def __eq__(self, other):
-        if isinstance(other, State):
-            return self.name == other.name and self.description == other.description and self.tags == other.tags
-        return False
-
-    def __hash__(self):
-        return hash((self.name, self.description, tuple(self.tags)))
-
-
-@dataclass
-class StateSpace:
-
-    states: Set[State]
-
-    def __str__(self):
-        return " ".join([str(state) for state in self.states])
-
-    def filter_states(self, condition: Callable[State, bool]) -> "StateSpace":
-        filtered_states = {state for state in self.states if condition(state)}
-        return StateSpace(filtered_states)
-
-
-@dataclass
-class MessageProfile:
-
-    message: AgentMessage
-    cost: float
-    duration: float
-    states: Set[State]  # unorddered collection of states
-
-    def __str__(self):
-        repr = f"Cost: {self.cost}\tDuration: {self.duration}\t"
-        for state in self.states:
-            repr += str(state) + " "
-        return repr
-
-
 class AppErrorMessage(Message):
     """An error message for the app."""
 
     def __init__(self, message: str) -> None:
         self.message = message
         super().__init__()
-
-
-class Profiler:
-
-    DEFAULT_STATE_SPACE = StateSpace(
-        states={
-            State(
-                name="USER-REQUEST",
-                description="The message shows the *user* requesting a task that needs to be completed",
-                tags=["user"],
-            ),
-            State(
-                name="CODING",
-                description="The message shows the assistant writing python or shell code to solve a problem. IE the message contains code blocks. This code does not apply to markdown code blocks",
-                tags=["assistant"],
-            ),
-            State(
-                name="PLANNING",
-                description="The message shows that the agent is create a step by step plan to accomplish some task.",
-                tags=["assistant"],
-            ),
-            State(
-                name="ANALYSING-RESULTS",
-                description="The assistant's message is reflecting on results obtained so far",
-                tags=["assistant"],
-            ),
-            State(
-                name="CODE-EXECUTION",
-                description="The user shared results of code execution, e.g., results, logs, error trace",
-                tags=["user"],
-            ),
-            State(
-                name="CODE-EXECUTION-ERROR",
-                description="The user shared results of code execution and they show an error in execution",
-                tags=["user"],
-            ),
-            State(
-                name="CODE-EXECUTION-SUCCESS",
-                description="The user shared results of code execution and they show a successful execution",
-                tags=["user"],
-            ),
-            State(
-                name="CODING-TOOL-USE",
-                description="The message contains a code block and the code uses method from the `functions` module eg indicated by presence of `from functions import....`",
-                tags=["assistant"],
-            ),
-            State(
-                name="ASKING-FOR-INFO",
-                description="The assistant is asking a question",
-                tags=["assistant"],
-            ),
-            State(
-                name="SUMMARIZING",
-                description="The assistant is synthesizing/summarizing information gathered so far",
-                tags=["assistant"],
-            ),
-            State(
-                name="TERMINATE", description="The agent's message contains the word 'TERMINATE'", tags=["assistant"]
-            ),
-            State(name="EMPTY", description="The message is empty", tags=["user"]),
-            State(
-                name="UNDEFINED",
-                description="Use this code when the message does not fit any of the other codes",
-                tags=["user", "assistant"],
-            ),
-        }
-    )
-
-    def __init__(self, state_space: StateSpace = None):
-        self.state_space = state_space or self.DEFAULT_STATE_SPACE
-
-    def profile_message(self, message: AgentMessage) -> MessageProfile:
-
-        def role_in_tags(state: State) -> bool:
-            # if state has no tags, the state applies to all roles
-            if state.tags is None:
-                return True
-            return message.role in state.tags
-
-        state_space = self.state_space.filter_states(condition=role_in_tags)
-        state_space_str = ""
-
-        for state in state_space.states:
-            state_space_str += f"{state.name}: {state.description}" + "\n"
-
-        prompt = f"""Which of the following codes apply to the message:
-List of codes:
-{state_space_str}
-
-Message
-    role: "{message.role}"
-    content: "{message.content}"
-
-Only respond with codes that apply. Codes should be separated by commas.
-    """
-        client = autogen.OpenAIWrapper(**LLM_CONFIG)
-        response = client.create(messages=[{"role": "user", "content": prompt}])
-        response = client.extract_text_or_completion_object(response)[0]
-
-        extracted_states_names = response.split(",")
-        extracted_states = []
-        for state_name in extracted_states_names:
-            extracted_states.append(State(name=state_name, description="", tags=[]))
-
-        message_profile = MessageProfile(cost=0.0, duration=0.0, states=extracted_states, message=message)
-
-        return message_profile
-
-
-@dataclass
-class ChatProfile:
-
-    num_messages: int
-    message_profiles: List[MessageProfile]  # ordered collection of message profiles
-
-    def __str__(self):
-        repr = f"Num messages: {self.num_messages}" + "\n"
-        for message_profile in self.message_profiles:
-            repr += str(message_profile) + "\n"
-        return repr
-
-
-class ProfileNode(Static):
-
-    message_profile: MessageProfile
-
-    DEFAULT_CSS = """
-    ProfileNode Markdown {
-        border: solid $primary;
-        padding: 1;
-    }
-"""
-
-    def compose(self) -> ComposeResult:
-        states = self.message_profile.states
-
-        def state_name_comparator(x: State, y: State):
-            return x.name < y.name
-
-        states.sort(key=functools.cmp_to_key(state_name_comparator))
-
-        state_display_str = " ".join([str(state) for state in states])
-
-        with Collapsible(collapsed=True, title=state_display_str):
-            yield Static(str(self.message_profile))
-            yield Markdown(str(self.message_profile.message))
-
-
-class ProfileDiagram(ScrollableContainer):
-
-    chat_profile: ChatProfile = reactive(None, recompose=True)
-
-    def compose(self) -> ComposeResult:
-
-        if self.chat_profile is None:
-            yield Label("Profiling...")
-            yield LoadingIndicator()
-            return
-
-        # if self.chat_profile.num_messages == 0:
-        #     yield LoadingIndicator()
-        #     return
-        num_messages = self.chat_profile.num_messages
-        yield Label(f"Number of messages: {num_messages}", classes="heading")
-        for message_profile in self.chat_profile.message_profiles:
-            node = ProfileNode()
-            node.message_profile = message_profile
-            yield node
-
-
-class ProfilerContainer(Container):
-
-    root_id = None
-    chat_history = reactive(None)
-    profile_diagram = None
-
-    def on_mount(self) -> None:
-        # self.update_chat_history()
-        # self.start_profiling()
-        self.set_interval(1, self.update_chat_history)
-
-    def update_chat_history(self) -> None:
-        self.chat_history = fetch_chat_history(self.root_id)
-
-    def watch_chat_history(self, new_chat_history) -> None:
-        if new_chat_history is None:
-            return
-
-        self.start_profiling()
-
-    @work(thread=True, exclusive=True)
-    async def start_profiling(self):
-        chat_profile = await self.profile_chat()
-        if self.profile_diagram is None:
-            self.profile_diagram = ProfileDiagram()
-        self.profile_diagram.chat_profile = chat_profile
-
-    async def profile_chat(self) -> ChatProfile:
-        profiler = Profiler()
-        message_profile_list = []
-        for message in self.chat_history:
-            _message = AgentMessage(role=message["role"], content=message["content"])
-            msg_profile = profiler.profile_message(_message)
-            message_profile_list.append(msg_profile)
-
-        chat_profile = ChatProfile(num_messages=len(self.chat_history), message_profiles=message_profile_list)
-
-        return chat_profile
-
-    def compose(self):
-        if self.profile_diagram is None:
-            self.profile_diagram = ProfileDiagram()
-        yield self.profile_diagram
-
-
-class ChatScreen(ModalScreen):
-    """A screen that displays a chat history"""
-
-    root_msg_id = 0
-    BINDINGS = [("escape", "app.pop_screen", "Pop screen")]
-
-    def compose(self) -> ComposeResult:
-        history = fetch_chat_history(self.root_msg_id)
-        with Grid(id="chat-screen"):
-
-            with Container(id="chat-screen-header"):
-                yield Label(f"Monitoring 🧵 Thread: {self.root_msg_id}", classes="heading")
-
-            with TabbedContent("Overview", "Details", id="chat-screen-tabs"):
-                profiler = ProfilerContainer(id="chat-profiler")
-                profiler.root_id = self.root_msg_id
-                profiler.chat_history = history
-
-                yield profiler
-
-                with ScrollableContainer(id="chat-screen-contents"):
-                    for msg in history:
-                        if msg["role"] == "assistant":
-                            msg_class = "assistant-message"
-                        if msg["role"] == "user":
-                            msg_class = "user-message"
-                        yield Markdown(f"{msg['role']}:\n{msg['content']}", classes=msg_class + " message")
-
-            with Horizontal(id="chat-screen-footer"):
-                yield Button("Learn New Tool", variant="error", id="learn")
-
-    @on(Button.Pressed, "#learn")
-    def learn(self) -> None:
-        learning_screen = LearningScreen()
-        learning_screen.root_msg_id = self.root_msg_id
-        self.app.push_screen(learning_screen)
-
-
-class LearningScreen(ModalScreen):
-
-    BINDINGS = [("escape", "app.pop_screen", "Pop screen")]
-
-    root_msg_id = None
-
-    def compose(self) -> ComposeResult:
-        with Grid(id="learning-screen"):
-            yield Horizontal(Label("Interactive Tool Learning", classes="heading"), id="learning-screen-header")
-            yield ScrollableContainer(
-                TextArea.code_editor(
-                    f"""
-                    # Learning a function for {self.root_msg_id}
-                    """,
-                    language="python",
-                ),
-                id="learning-screen-contents",
-            )
-            with Horizontal(id="learning-screen-footer"):
-                # yield Button("Start", variant="error", id="start-learning")
-                yield Button("Save", variant="primary", id="save")
-
-    def on_mount(self) -> None:
-        self.start_learning()
-
-    @on(Button.Pressed, "#save")
-    def save(self) -> None:
-        widget = self.query_one("#learning-screen-contents > TextArea", TextArea)
-        code = widget.text
-        name = code.split("\n")[0][1:]
-
-        tool = Tool(name, code)
-        try:
-            tool.validate_tool()
-            APP_CONFIG.update_tool(tool)
-            self.app.pop_screen()
-            self.app.push_screen(NotificationScreen(message="Tool saved successfully"))
-
-        except InvalidToolError as e:
-            error_message = f"{e}"
-            self.post_message(AppErrorMessage(error_message))
-            return
-
-        except ToolUpdateError as e:
-            error_message = f"{e}"
-            self.post_message(AppErrorMessage(error_message))
-            return
-
-    @work(thread=True)
-    async def start_learning(self) -> None:
-        widget = self.query_one("#learning-screen-contents > TextArea", TextArea)
-        widget.text = "# Learning..."
-
-        history = await a_fetch_chat_history(self.root_msg_id)
-        name, code = learn_tool_from_history(history)
-
-        widget.text = "#" + name + "\n" + code
-
-
-def learn_tool_from_history(history: List[Dict[str, str]]) -> str:
-
-    # return "hola"
-
-    markdown = ""
-    for msg in history:
-        markdown += f"{msg['role']}: {msg['content']}\n"
-
-    agent = ConversableAgent(
-        "learning_assistant",
-        llm_config=LLM_CONFIG,
-        system_message="""You are a helpful assistant that for the given chat
-history can return a standalone, documented python function.
-
-Try to extract a most general version of the function based on the chat history.
-That can be reused in the future for similar tasks. Eg do not use hardcoded arguments.
-Instead make them function parameters.
-
-The chat history contains a task the agents were trying to accomplish.
-Analyze the following chat history to assess if the task was completed,
-and if it was return the python function that would accomplish the task.
-        """,
-    )
-    messages = [
-        {
-            "role": "user",
-            "content": f"""The chat history is
-
-            {markdown}
-
-            Only generate a single python function in code blocks and nothing else.
-            Make sure all imports are inside the function.
-            Both ast.FunctionDef, ast.AsyncFunctionDef are acceptable.
-
-            Function signature should be annotated properly.
-            Function should return a string as the final result.
-            """,
-        }
-    ]
-    reply = agent.generate_reply(messages)
-    from autogen.code_utils import extract_code
-
-    # extract a code block from the reply
-    code_blocks = extract_code(reply)
-    lang, code = code_blocks[0]
-
-    messages.append({"role": "assistant", "content": code})
-
-    messages.append(
-        {
-            "role": "user",
-            "content": """suggest a max two word english phrase that is a friendly
-          display name for the function. Only reply with the name in a code block.
-          no need to use an quotes or code blocks. Just two words.""",
-        }
-    )
-
-    name = agent.generate_reply(messages)
-
-    return name, code
 
 
 def generate_response_process(msg_idx: int):
